fix(agent): address review — poll budget window and Windows test portability
Two blockers from review, both confirmed against the code:

1. The client poll budget did not cover the server's full request lifetime.
   The pending and running windows are sequential, not overlapping: pending is
   measured from CreatedAt (30s to claim), running from RunStartedAt, which
   PopPending sets at claim time. Heartbeat pickup therefore does NOT consume
   the 60s running window, as the previous comment claimed. A request claimed
   at ~29s that then runs hermes' ~40s discovery reports at ~69s — past the 60s
   client budget, while the server still holds the record open. The user got a
   generic client-side "timed out" instead of the daemon's reason, which is the
   exact failure mode this PR exists to remove.

   POLL_TIMEOUT_MS is now derived from the two server constants plus slack, so
   the invariant is visible rather than asserted in prose. Letting the server's
   own timeout win is also strictly better copy ("daemon did not respond within
   30 seconds" beats "model discovery timed out"). Added two fake-timer tests:
   one for a late claim plus a full discovery, one proving the poll still
   terminates once both server windows have closed.

2. The new backend regression tests broke `go test ./pkg/agent` on Windows.
   They write and execute `#!/bin/sh` fixtures, but the file carried no build
   tag, and the Windows writeTestExecutable can only drop bytes on disk. Split
   along the same line as codebuddy_discovery_fallback_test.go: process-based
   cases move to hermes_model_discovery_unix_test.go (//go:build !windows), and
   the assertions needing no subprocess — hint wording, classification
   inertness, the timeout invariant — stay cross-platform against a literal of
   the ACP error string.

Also corrects the Go timeout invariant itself, which had encoded the same wrong
model of the running window, and refreshes two daemon.go comments that still
described discovery as capped at ~15s and as uniformly returning empty lists on
failure.

Verified: GOOS=windows go test -c builds, with the shell-fixture tests excluded
from the Windows binary and all four portable tests present. Both new core tests
fail against the previous 60s budget.
